python/requirements.txt
Flask>=2.2.0,<2.3.0
PyJWT>=2.6.0,<=2.7.0
PyMySQL>=1.0.0,<1.1.0
SQLAlchemy>=2.0.0,<2.1.0
gunicorn>=20.1.0,<20.2.0
